// x-pack/platform/plugins/shared/fleet/server/constants/api_privileges.d.ts
export declare const FLEET_API_PRIVILEGES: {
    FLEET: {
        READ: string;
        ALL: string;
    };
    AGENTS: {
        READ: string;
        ALL: string;
    };
    AGENT_POLICIES: {
        READ: string;
        ALL: string;
    };
    SETTINGS: {
        READ: string;
        ALL: string;
    };
    GENERATE_REPORTS: {
        ALL: string;
    };
    INTEGRATIONS: {
        READ: string;
        ALL: string;
    };
    SETUP: string;
};